services/url_encoder/src/requirements.txt
grpcio
grpcio-tools

psycopg[binary]